Unify enabling and disabling
