+++ distribution/lib/Standard/Base/0.0.0-dev/src/Internal/Context_Helpers.enso
@@ -14,5 +14,4 @@ Integer.from (that:Context) = case that of
 
 type Context_Helpers
     is_enabled c env = @Builtin_Method "Context.is_enabled_builtin"
-    with_enabled_context_builtin context environment ~action = @Builtin_Method "Runtime.with_enabled_context_builtin"
-    with_disabled_context_builtin context environment ~action = @Builtin_Method "Runtime.with_disabled_context_builtin"
+    with_context context enable ~action = @Builtin_Method "Runtime.with_context_builtin"
